Register functionality Added

+++ base/views.py
@@ -1,14 +1,28 @@
+from django.http import HttpResponse
 from os import name
 from .models import Room, Topic
 from django.shortcuts import render, redirect
-from .forms import RoomForm
+from .forms import RoomForm, UserForm
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 
+def registerUser(request):
+    if request.method == "POST":
+        form = UserForm(request.POST)
+        if form.is_valid:
+            form.save()
+            user = User.objects.get(username = request.POST.get('username'))
+            print(user)
+            login(request, user)
+            return redirect('home')
+    form = UserForm()
+    action = "register"
+    context = {"action": action, "form": form}
+    return render(request, 'base/login_register.html', context)
 
-# Create your views here.
 
 def logoutPage(request):
     #user = request.user
@@ -17,6 +31,9 @@ def logoutPage(request):
     return redirect('home')
 
 def loginPage(request):
+
+    if request.user.is_authenticated :
+        return redirect('home')
 
     if request.method == 'POST':
         username = request.POST.get('username')
@@ -33,8 +50,9 @@ def loginPage(request):
             return redirect('home')
         else:
             messages.error(request, "Username and Password not matched ")
-
-    return render(request, 'base/login_register.html')
+    action="login"
+    context = {"action": action}
+    return render(request, 'base/login_register.html', context)
 
 def home(request,):
     query = request.GET.get('q')
@@ -54,6 +72,7 @@ def room(request, pk):
     context = {"room": room}
     return render(request, "base/room.html", context= context)
 
+@login_required
 def add_room(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -64,9 +83,12 @@ def add_room(request):
     context = {"form": form}
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url= '/login')
 def edit_room(request, pk):
+    
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse("You are not authorized to perform this action !")
     form = RoomForm(instance=room)
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -77,8 +99,11 @@ def edit_room(request, pk):
     context = {"form": form}
     return render(request, 'base/room_form.html', context)
 
+@login_required
 def delete_room(request, pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse("You are not authorized to perform this action !")
     if request.method == 'POST':
         room.delete()
         return redirect('home')
